refactor(pipeline): extract helpers_test.go from engine_test.go [#284]
Move shared test helpers (mock runners, setup functions, utilities) from
engine_test.go into a dedicated helpers_test.go to improve readability
and reduce the size of the 10k-line test file.

Extracted: flexMockRunner, setupEngine, initGitRepo, phaseNames,
setupReviewEngine, chunkMockRunner, blockingChunkRunner,
capturingChunkRunner, funcRunner.
