+++ internal/pipeline/engine_test.go
@@ -7,7 +7,6 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"strings"
 	"sync"
@@ -17,87 +16,6 @@ import (
 	"github.com/decko/soda/internal/runner"
 	"github.com/decko/soda/schemas"
 )
-
-// flexMockRunner returns per-call responses, allowing multi-call test scenarios
-// (e.g., fail twice then succeed).
-type flexMockRunner struct {
-	mu        sync.Mutex
-	responses map[string][]flexResponse
-	calls     []runner.RunOpts
-	counters  map[string]int
-}
-
-type flexResponse struct {
-	result *runner.RunResult
-	err    error
-}
-
-func (f *flexMockRunner) Run(ctx context.Context, opts runner.RunOpts) (*runner.RunResult, error) {
-	f.mu.Lock()
-	defer f.mu.Unlock()
-
-	f.calls = append(f.calls, opts)
-	if f.counters == nil {
-		f.counters = make(map[string]int)
-	}
-	idx := f.counters[opts.Phase]
-	f.counters[opts.Phase]++
-	resps, ok := f.responses[opts.Phase]
-	if !ok || idx >= len(resps) {
-		return nil, fmt.Errorf("flexmock: no response for phase %q call %d", opts.Phase, idx)
-	}
-	resp := resps[idx]
-	return resp.result, resp.err
-}
-
-// setupEngine creates temp directories, writes minimal prompt templates,
-// creates State, and returns an Engine + State ready for testing.
-func setupEngine(t *testing.T, phases []PhaseConfig, r runner.Runner, opts ...func(*EngineConfig)) (*Engine, *State) {
-	t.Helper()
-
-	stateDir := t.TempDir()
-	promptDir := t.TempDir()
-	workDir := t.TempDir()
-
-	// Write a minimal prompt template for each phase.
-	for _, p := range phases {
-		tmplPath := filepath.Join(promptDir, p.Prompt)
-		if err := os.MkdirAll(filepath.Dir(tmplPath), 0755); err != nil {
-			t.Fatalf("mkdir for prompt %s: %v", p.Prompt, err)
-		}
-		content := fmt.Sprintf("Phase: %s\nTicket: {{.Ticket.Key}}\n", p.Name)
-		if err := os.WriteFile(tmplPath, []byte(content), 0644); err != nil {
-			t.Fatalf("write prompt %s: %v", p.Prompt, err)
-		}
-	}
-
-	state, err := LoadOrCreate(stateDir, "TEST-1")
-	if err != nil {
-		t.Fatalf("LoadOrCreate: %v", err)
-	}
-
-	pipeline := &PhasePipeline{Phases: phases}
-	loader := NewPromptLoader(promptDir)
-
-	cfg := EngineConfig{
-		Pipeline:   pipeline,
-		Loader:     loader,
-		Ticket:     TicketData{Key: "TEST-1", Summary: "Test ticket"},
-		Model:      "test-model",
-		WorkDir:    workDir,
-		MaxCostUSD: 0, // no budget limit by default
-		Mode:       Autonomous,
-		SleepFunc:  func(time.Duration) {}, // no-op sleep for tests
-		JitterFunc: func(time.Duration) time.Duration { return 0 },
-	}
-
-	for _, opt := range opts {
-		opt(&cfg)
-	}
-
-	engine := NewEngine(r, state, cfg)
-	return engine, state
-}
 
 func TestEngine_HappyPathAllPhasesComplete(t *testing.T) {
 	phases := []PhaseConfig{
@@ -1971,23 +1889,6 @@ func TestEngine_ResumeRerunsCompletedPhase(t *testing.T) {
 	}
 }
 
-// initGitRepo creates a bare-minimum git repo in dir with one commit on "main".
-func initGitRepo(t *testing.T, dir string) {
-	t.Helper()
-	for _, args := range [][]string{
-		{"git", "init", "--initial-branch=main"},
-		{"git", "config", "user.email", "test@test.com"},
-		{"git", "config", "user.name", "test"},
-		{"git", "commit", "--allow-empty", "-m", "init"},
-	} {
-		cmd := exec.Command(args[0], args[1:]...)
-		cmd.Dir = dir
-		if out, err := cmd.CombinedOutput(); err != nil {
-			t.Fatalf("%v: %s: %v", args, out, err)
-		}
-	}
-}
-
 func TestEngine_WorktreeCreatedBeforeFirstPhase(t *testing.T) {
 	phases := []PhaseConfig{
 		{
@@ -3525,79 +3426,6 @@ func TestEngine_PromptLoadedFallbackEvent(t *testing.T) {
 	if !state.IsCompleted("triage") {
 		t.Error("triage should be completed after fallback")
 	}
-}
-
-// phaseNames extracts phase names from runner calls for test error messages.
-func phaseNames(calls []runner.RunOpts) []string {
-	names := make([]string, len(calls))
-	for i, c := range calls {
-		names[i] = c.Phase
-	}
-	return names
-}
-
-// setupReviewEngine creates temp directories, writes prompt templates for
-// reviewer-specific prompts, creates State, and returns an Engine + State
-// for testing parallel-review phases.
-func setupReviewEngine(t *testing.T, phases []PhaseConfig, r runner.Runner, opts ...func(*EngineConfig)) (*Engine, *State) {
-	t.Helper()
-
-	stateDir := t.TempDir()
-	promptDir := t.TempDir()
-	workDir := t.TempDir()
-
-	// Write prompt templates for regular phases.
-	for _, p := range phases {
-		if p.Prompt != "" {
-			tmplPath := filepath.Join(promptDir, p.Prompt)
-			if err := os.MkdirAll(filepath.Dir(tmplPath), 0755); err != nil {
-				t.Fatalf("mkdir for prompt %s: %v", p.Prompt, err)
-			}
-			content := fmt.Sprintf("Phase: %s\nTicket: {{.Ticket.Key}}\n", p.Name)
-			if err := os.WriteFile(tmplPath, []byte(content), 0644); err != nil {
-				t.Fatalf("write prompt %s: %v", p.Prompt, err)
-			}
-		}
-
-		// Write reviewer prompt templates for parallel-review phases.
-		for _, reviewer := range p.Reviewers {
-			tmplPath := filepath.Join(promptDir, reviewer.Prompt)
-			if err := os.MkdirAll(filepath.Dir(tmplPath), 0755); err != nil {
-				t.Fatalf("mkdir for reviewer prompt %s: %v", reviewer.Prompt, err)
-			}
-			content := fmt.Sprintf("Reviewer: %s\nFocus: %s\nTicket: {{.Ticket.Key}}\n", reviewer.Name, reviewer.Focus)
-			if err := os.WriteFile(tmplPath, []byte(content), 0644); err != nil {
-				t.Fatalf("write reviewer prompt %s: %v", reviewer.Prompt, err)
-			}
-		}
-	}
-
-	state, err := LoadOrCreate(stateDir, "TEST-1")
-	if err != nil {
-		t.Fatalf("LoadOrCreate: %v", err)
-	}
-
-	pipeline := &PhasePipeline{Phases: phases}
-	loader := NewPromptLoader(promptDir)
-
-	cfg := EngineConfig{
-		Pipeline:   pipeline,
-		Loader:     loader,
-		Ticket:     TicketData{Key: "TEST-1", Summary: "Test ticket"},
-		Model:      "test-model",
-		WorkDir:    workDir,
-		MaxCostUSD: 0,
-		Mode:       Autonomous,
-		SleepFunc:  func(time.Duration) {},
-		JitterFunc: func(time.Duration) time.Duration { return 0 },
-	}
-
-	for _, opt := range opts {
-		opt(&cfg)
-	}
-
-	engine := NewEngine(r, state, cfg)
-	return engine, state
 }
 
 func TestEngine_ParallelReview_HappyPath(t *testing.T) {
@@ -5834,22 +5662,6 @@ func TestEngine_PhaseLifecycleEvents(t *testing.T) {
 	})
 }
 
-// chunkMockRunner is a flexMockRunner that invokes OnChunk before returning.
-type chunkMockRunner struct {
-	flexMockRunner
-	chunks map[string][]string // phase name → lines to emit via OnChunk
-}
-
-func (c *chunkMockRunner) Run(ctx context.Context, opts runner.RunOpts) (*runner.RunResult, error) {
-	// Emit chunks before returning the result.
-	if lines, ok := c.chunks[opts.Phase]; ok && opts.OnChunk != nil {
-		for _, line := range lines {
-			opts.OnChunk(line)
-		}
-	}
-	return c.flexMockRunner.Run(ctx, opts)
-}
-
 func TestEngine_OutputChunkEvents(t *testing.T) {
 	phases := []PhaseConfig{
 		{
@@ -6112,21 +5924,6 @@ func TestEngine_PauseBlocksOutputStreaming(t *testing.T) {
 	}
 }
 
-// blockingChunkRunner is a test runner that calls OnChunk for each line.
-type blockingChunkRunner struct {
-	result *runner.RunResult
-	chunks []string
-}
-
-func (b *blockingChunkRunner) Run(ctx context.Context, opts runner.RunOpts) (*runner.RunResult, error) {
-	for _, line := range b.chunks {
-		if opts.OnChunk != nil {
-			opts.OnChunk(line)
-		}
-	}
-	return b.result, nil
-}
-
 func TestEngine_NilPauseSignalNoOp(t *testing.T) {
 	phases := []PhaseConfig{
 		{
@@ -6196,19 +5993,6 @@ func TestEngine_OnChunkPassedToRunner(t *testing.T) {
 	if capturedOnChunk == nil {
 		t.Fatal("expected OnChunk to be set in RunOpts")
 	}
-}
-
-// capturingChunkRunner captures the OnChunk function from RunOpts.
-type capturingChunkRunner struct {
-	result         *runner.RunResult
-	captureOnChunk func(func(string))
-}
-
-func (c *capturingChunkRunner) Run(ctx context.Context, opts runner.RunOpts) (*runner.RunResult, error) {
-	if c.captureOnChunk != nil {
-		c.captureOnChunk(opts.OnChunk)
-	}
-	return c.result, nil
 }
 
 func TestEngine_DrainPauseSignalUnpausesOnClose(t *testing.T) {
@@ -6346,15 +6130,6 @@ func TestEngine_OnChunkContextCancel(t *testing.T) {
 	case <-time.After(5 * time.Second):
 		t.Fatal("engine deadlocked — OnChunk did not unblock on context cancel")
 	}
-}
-
-// funcRunner is a test runner that calls a function.
-type funcRunner struct {
-	fn func(context.Context, runner.RunOpts) (*runner.RunResult, error)
-}
-
-func (f *funcRunner) Run(ctx context.Context, opts runner.RunOpts) (*runner.RunResult, error) {
-	return f.fn(ctx, opts)
 }
 
 func TestEngine_OutputChunkNotLoggedToFile(t *testing.T) {
